102 Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-int a, b, c, d;
+long a, b, c, d;
 a = 0;
 b = 1;
 for (c = 0; c < 50; c++)
@@ -16,9 +16,9 @@ d = a + b;
 a = b;
 b = d;
 if (c != 49)
-printf("%d, ", d);
+printf("%ld, ", d);
 else
-printf("%d\n", d);
+printf("%ld\n", d);
 }
 return (0);
 }
